+++ lib/nice_partials/partial.rb
@@ -32,6 +32,23 @@ module NicePartials
       class_eval &block
     end
 
+    # `translate` is a shorthand to set `content_for` with content that's run through
+    # the view's `translate`/`t` context.
+    #
+    #   partial.t :title                       # => partial.content_for :title, t(".title")
+    #   partial.t title: :section              # => partial.content_for :title, t(".section")
+    #   partial.t title: "some.custom.key"     # => partial.content_for :title, t("some.custom.key")
+    #   partial.t :description, title: :header # Mixing is supported too.
+    #
+    # Note that `partial.t "some.custom.key"` can't derive a `content_for` name, so an explicit
+    # name must be provided e.g. `partial.t title: "some.custom.key"`.
+    def translate(*names, **renames)
+      names.chain(renames).each do |name, key = name|
+        content_for name, @view_context.t(key.is_a?(String) ? key : ".#{key}")
+      end
+    end
+    alias t translate
+
     # Allows an inner partial to copy content from an outer partial.
     #
     # Additionally a hash of keys to rename in the new partial context can be passed.
@@ -56,7 +73,7 @@ module NicePartials
     # and lets you pass arguments into it, like so:
     #
     #   # Here we store a block with some eventual content…
-    #   <% partial.title.store(&:h1) %>
+    #   <% partial.title { |tag| tag.h1 } %>
     #
     #  # …which we can then yield into with some predefined options later.
     #  <%= partial.title.yield tag.with_options(class: "text-bold") %>
@@ -70,7 +87,7 @@ module NicePartials
     alias content_for? section?
 
     def content_for(...)
-      section(...).to_s
+      section(...)&.to_s
     end
 
     def capture(*arguments, &block)