+++ test/nice_partials/partial_test.rb
@@ -62,17 +62,16 @@ class NicePartials::PartialTest < NicePartials::Test
     assert_equal %(<p class="post-title">content</p>),   partial.title.p
     assert_equal %(<h2 class="post-title">content</h2>), partial.title.h2
 
-    unless defined?(AttributesAndTokenLists)
-      assert_equal %(<h2 class="">content</h2>),                 partial.title.h2(class: { "text-m4": false })
-      assert_equal %(<h2 class="text-m4">contentaddendum</h2>),  partial.title.h2("addendum", class: "text-m4")
-      assert_equal %(<h2 class="some-class">contentblabla</h2>), partial.title.h2("blabla", class: "some-class")
-    end
+    assert_equal %(<h2 class="">content</h2>),                 partial.title.h2(class: { "text-m4": false })
+    assert_equal %(<h2 class="text-m4">contentaddendum</h2>),  partial.title.h2("addendum", class: "text-m4")
+    assert_equal %(<h2 class="some-class">contentblabla</h2>), partial.title.h2("blabla", class: "some-class")
   end
 
   test "content_for returns content itself and not section object" do
     partial = new_partial
     partial.body "some content"
-    partial.content_for :body, ", yet more"
+
+    assert_nil partial.content_for(:body, ", yet more"), "content_for must return nil when writing content"
 
     assert_equal "some content, yet more", partial.content_for(:body)
     assert_equal "some content, yet more", partial.body.to_s