+++ test/renderer/translation_test.rb
@@ -4,10 +4,12 @@ module Renderer; end
 
 class Renderer::TranslationTest < NicePartials::Test
   setup do
-    I18n.backend.store_translations "en", { translations: {
+    I18n.backend.store_translations "en", translations: {
       translated: { message: "message" },
-      nice_partials_translated: { message: "nice_partials" }
-    } }
+      nice_partials_translated: { message: "nice_partials" },
+      t: { title: "title key content", header: "header key content" },
+    }
+    I18n.backend.store_translations "en", custom: { key: "custom key content" }
   end
 
   teardown { I18n.reload! }
@@ -22,5 +24,14 @@ class Renderer::TranslationTest < NicePartials::Test
     render "translations/nice_partials_translated"
 
     assert_text "nice_partials"
+  end
+
+  test "translate method" do
+    partial = nil
+    render("translations/t") { partial = _1 }
+
+    assert_equal "title key content",  partial.title.to_s
+    assert_equal "header key content", partial.description.to_s
+    assert_equal "custom key content", partial.byline.to_s
   end
 end