+++ test/test_helper.rb
@@ -1,10 +1,3 @@
-require "active_support"
-require "active_support/testing/autorun"
-require "action_controller"
-require "action_view"
-require "action_view/test_case"
-require "capybara/minitest"
-
 ENV["RAILS_ENV"] = "test"
 require "rails"
 require "rails/test_help"
@@ -15,6 +8,7 @@ class TestApp < Rails::Application
   secrets.secret_key_base = "secret_key_base"
 end
 
+require "capybara/minitest"
 require "view_component"
 require "nice_partials"